Move postgresql tests to their own file
